Deduplicate StrEnum
A single `StrEnum` class is defined and imported by `Order` and `Asset`
classes.

Resolves #788
